+++ deployment/v2_0_0/changesets/apply_executor_config_test.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/BurntSushi/toml"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 
 	chainsel "github.com/smartcontractkit/chain-selectors"
@@ -18,8 +17,6 @@ import (
 	"github.com/smartcontractkit/chainlink-deployments-framework/datastore"
 	"github.com/smartcontractkit/chainlink-deployments-framework/deployment"
 	"github.com/smartcontractkit/chainlink-deployments-framework/operations"
-	jobpb "github.com/smartcontractkit/chainlink-protos/job-distributor/v1/job"
-	nodev1 "github.com/smartcontractkit/chainlink-protos/job-distributor/v1/node"
 
 	"github.com/smartcontractkit/chainlink-ccip/deployment/v2_0_0/adapters"
 	"github.com/smartcontractkit/chainlink-ccip/deployment/v2_0_0/changesets"
@@ -228,23 +225,11 @@ func TestApplyExecutorConfig_HappyPathBuildsJobSpecs(t *testing.T) {
 	topo := newMinimalTopology([]string{"nop1", "nop2"}, "pool1", shared.NOPModeStandalone)
 	env := newTestExecutorEnv(t, []uint64{sel1})
 	mockJD := mocks.NewMockClient(t)
-	mockJD.EXPECT().ListNodes(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodesResponse{Nodes: []*nodev1.Node{
-			{Id: "node-1", Name: "nop1"},
-			{Id: "node-2", Name: "nop2"},
-		}}, nil,
-	)
-	mockJD.EXPECT().ListNodeChainConfigs(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodeChainConfigsResponse{ChainConfigs: []*nodev1.ChainConfig{
-			chainConfig("node-1", "90000001"),
-			chainConfig("node-2", "90000001"),
-		}}, nil,
-	)
-	mockJD.EXPECT().ProposeJob(mock.Anything, mock.Anything).Return(
-		&jobpb.ProposeJobResponse{Proposal: &jobpb.Proposal{Id: "job-1"}}, nil,
-	)
 	env.Offchain = mockJD
-	env.NodeIDs = []string{"node-1", "node-2"}
+	env.NodeIDs = expectJDInteractions(t, mockJD, []mockJDNode{
+		{nodeID: "node-1", nopAlias: "nop1", chainIDs: []string{"90000001"}},
+		{nodeID: "node-2", nopAlias: "nop2", chainIDs: []string{"90000001"}},
+	}, false)
 
 	cs := changesets.ApplyExecutorConfig(registry, newTestChainFamilyRegistry())
 	output, err := cs.Apply(env, changesets.ApplyExecutorConfigInput{
@@ -336,21 +321,10 @@ func TestApplyExecutorConfig_UsesAllDeployedChains(t *testing.T) {
 	topo := newMinimalTopology([]string{"nop1"}, "pool1", shared.NOPModeStandalone)
 	env := newTestExecutorEnv(t, []uint64{sel1, sel2})
 	mockJD := mocks.NewMockClient(t)
-	mockJD.EXPECT().ListNodes(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodesResponse{Nodes: []*nodev1.Node{
-			{Id: "node-1", Name: "nop1"},
-		}}, nil,
-	)
-	mockJD.EXPECT().ListNodeChainConfigs(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodeChainConfigsResponse{ChainConfigs: []*nodev1.ChainConfig{
-			chainConfig("node-1", "90000001"),
-		}}, nil,
-	)
-	mockJD.EXPECT().ProposeJob(mock.Anything, mock.Anything).Return(
-		&jobpb.ProposeJobResponse{Proposal: &jobpb.Proposal{Id: "job-1"}}, nil,
-	)
 	env.Offchain = mockJD
-	env.NodeIDs = []string{"node-1"}
+	env.NodeIDs = expectJDInteractions(t, mockJD, []mockJDNode{
+		{nodeID: "node-1", nopAlias: "nop1", chainIDs: []string{"90000001"}},
+	}, false)
 
 	cs := changesets.ApplyExecutorConfig(registry, newTestChainFamilyRegistry())
 	output, err := cs.Apply(env, changesets.ApplyExecutorConfigInput{
@@ -383,23 +357,12 @@ func TestApplyExecutorConfig_TargetNOPsFiltersJobSpecs(t *testing.T) {
 	topo := newMinimalTopology([]string{"nop1", "nop2", "nop3"}, "pool1", shared.NOPModeStandalone)
 	env := newTestExecutorEnv(t, []uint64{sel1})
 	mockJD := mocks.NewMockClient(t)
-	mockJD.EXPECT().ListNodes(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodesResponse{Nodes: []*nodev1.Node{
-			{Id: "node-1", Name: "nop1"},
-			{Id: "node-2", Name: "nop2"},
-			{Id: "node-3", Name: "nop3"},
-		}}, nil,
-	)
-	mockJD.EXPECT().ListNodeChainConfigs(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodeChainConfigsResponse{ChainConfigs: []*nodev1.ChainConfig{
-			chainConfig("node-1", "90000001"),
-		}}, nil,
-	)
-	mockJD.EXPECT().ProposeJob(mock.Anything, mock.Anything).Return(
-		&jobpb.ProposeJobResponse{Proposal: &jobpb.Proposal{Id: "job-1"}}, nil,
-	)
 	env.Offchain = mockJD
-	env.NodeIDs = []string{"node-1", "node-2", "node-3"}
+	env.NodeIDs = expectJDInteractions(t, mockJD, []mockJDNode{
+		{nodeID: "node-1", nopAlias: "nop1", chainIDs: []string{"90000001"}},
+		{nodeID: "node-2", nopAlias: "nop2"},
+		{nodeID: "node-3", nopAlias: "nop3"},
+	}, false)
 
 	cs := changesets.ApplyExecutorConfig(registry, newTestChainFamilyRegistry())
 	output, err := cs.Apply(env, changesets.ApplyExecutorConfigInput{
@@ -522,27 +485,11 @@ func TestApplyExecutorConfig_PerChainNOPsIncludesAllChains(t *testing.T) {
 
 	env := newTestExecutorEnv(t, []uint64{sel1, sel2, sel3})
 	mockJD := mocks.NewMockClient(t)
-	mockJD.EXPECT().ListNodes(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodesResponse{Nodes: []*nodev1.Node{
-			{Id: "node-1", Name: "exec1"},
-			{Id: "node-2", Name: "exec2"},
-		}}, nil,
-	)
-	mockJD.EXPECT().ListNodeChainConfigs(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodeChainConfigsResponse{ChainConfigs: []*nodev1.ChainConfig{
-			chainConfig("node-1", "90000001"),
-			chainConfig("node-1", "90000002"),
-			chainConfig("node-1", "90000003"),
-			chainConfig("node-2", "90000001"),
-			chainConfig("node-2", "90000002"),
-			chainConfig("node-2", "90000003"),
-		}}, nil,
-	)
-	mockJD.EXPECT().ProposeJob(mock.Anything, mock.Anything).Return(
-		&jobpb.ProposeJobResponse{Proposal: &jobpb.Proposal{Id: "job-1"}}, nil,
-	)
 	env.Offchain = mockJD
-	env.NodeIDs = []string{"node-1", "node-2"}
+	env.NodeIDs = expectJDInteractions(t, mockJD, []mockJDNode{
+		{nodeID: "node-1", nopAlias: "exec1", chainIDs: []string{"90000001", "90000002", "90000003"}},
+		{nodeID: "node-2", nopAlias: "exec2", chainIDs: []string{"90000001", "90000002", "90000003"}},
+	}, false)
 	cs := changesets.ApplyExecutorConfig(registry, newTestChainFamilyRegistry())
 	output, err := cs.Apply(env, changesets.ApplyExecutorConfigInput{
 		Topology:          topo,
@@ -589,25 +536,11 @@ func TestApplyExecutorConfig_AllChainsNOPsGetAllChains(t *testing.T) {
 	)
 	env := newTestExecutorEnv(t, []uint64{sel1, sel2})
 	mockJD := mocks.NewMockClient(t)
-	mockJD.EXPECT().ListNodes(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodesResponse{Nodes: []*nodev1.Node{
-			{Id: "node-1", Name: "exec1"},
-			{Id: "node-2", Name: "exec2"},
-		}}, nil,
-	)
-	mockJD.EXPECT().ListNodeChainConfigs(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodeChainConfigsResponse{ChainConfigs: []*nodev1.ChainConfig{
-			chainConfig("node-1", "90000001"),
-			chainConfig("node-1", "90000002"),
-			chainConfig("node-2", "90000001"),
-			chainConfig("node-2", "90000002"),
-		}}, nil,
-	)
-	mockJD.EXPECT().ProposeJob(mock.Anything, mock.Anything).Return(
-		&jobpb.ProposeJobResponse{Proposal: &jobpb.Proposal{Id: "job-1"}}, nil,
-	)
 	env.Offchain = mockJD
-	env.NodeIDs = []string{"node-1", "node-2"}
+	env.NodeIDs = expectJDInteractions(t, mockJD, []mockJDNode{
+		{nodeID: "node-1", nopAlias: "exec1", chainIDs: []string{"90000001", "90000002"}},
+		{nodeID: "node-2", nopAlias: "exec2", chainIDs: []string{"90000001", "90000002"}},
+	}, false)
 
 	cs := changesets.ApplyExecutorConfig(registry, newTestChainFamilyRegistry())
 	output, err := cs.Apply(env, changesets.ApplyExecutorConfigInput{
@@ -656,25 +589,11 @@ func TestApplyExecutorConfig_ExecutorPoolFieldMatchesChainPool(t *testing.T) {
 
 	env := newTestExecutorEnv(t, []uint64{sel1, sel2})
 	mockJD := mocks.NewMockClient(t)
-	mockJD.EXPECT().ListNodes(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodesResponse{Nodes: []*nodev1.Node{
-			{Id: "node-1", Name: "exec1"},
-			{Id: "node-2", Name: "exec2"},
-		}}, nil,
-	)
-	mockJD.EXPECT().ListNodeChainConfigs(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodeChainConfigsResponse{ChainConfigs: []*nodev1.ChainConfig{
-			chainConfig("node-1", "90000001"),
-			chainConfig("node-1", "90000002"),
-			chainConfig("node-2", "90000001"),
-			chainConfig("node-2", "90000002"),
-		}}, nil,
-	)
-	mockJD.EXPECT().ProposeJob(mock.Anything, mock.Anything).Return(
-		&jobpb.ProposeJobResponse{Proposal: &jobpb.Proposal{Id: "job-1"}}, nil,
-	)
 	env.Offchain = mockJD
-	env.NodeIDs = []string{"node-1", "node-2"}
+	env.NodeIDs = expectJDInteractions(t, mockJD, []mockJDNode{
+		{nodeID: "node-1", nopAlias: "exec1", chainIDs: []string{"90000001", "90000002"}},
+		{nodeID: "node-2", nopAlias: "exec2", chainIDs: []string{"90000001", "90000002"}},
+	}, false)
 	cs := changesets.ApplyExecutorConfig(registry, newTestChainFamilyRegistry())
 	output, err := cs.Apply(env, changesets.ApplyExecutorConfigInput{
 		Topology:          topo,
